Move all parser related code to the parser module
